Move secrets logic to "ci" subcommand
This hacks around a little to provide backward compatibility to people
still used to run "fioctl secrets"
